fix privilege admin 3
