+++ client/js/secure.js
@@ -1,25 +1,39 @@
 import { displayModalUsers } from './modal-users-display.js'
 
-const checkUser = () => {
-    fetch(`http://localhost:3030/route-session/secure`, {'credentials': 'include',})
+export function checkUser() {
+    return fetch(`http://localhost:3030/route-session/secure`, {'credentials': 'include',})
     .then(res => res.json())
-    .then(res => {
-      if (typeof res === 'string') {
-        window.location = 'index.html'
-        return
-      }
-      if(res.username) {
-        document.getElementById('session-display-name').innerHTML = res.username
-        document.isAdminSecure = res.admin === 'true'
-        if(res.admin === 'true') {
-          document.getElementById('block-link-manage-admin').innerHTML = `
-          <a id="link-manage-users" href="#">Manage Users</a>
-          `
-          // display modal Manage Users
-          document.getElementById('link-manage-users').addEventListener('click', displayModalUsers)
-        }
-      }
-    })
-    }
+}
     
-checkUser()
+export const displayLinkManager = (res) => {
+
+  if (typeof res === 'string') {
+    window.location = 'index.html'
+    return
+  }
+  if(res.username) {
+    document.getElementById('session-display-name').innerHTML = res.username
+    document.isAdminSecure = res.admin === 'true'
+    if(res.admin === 'true') {
+      document.getElementById('block-link-manage-admin').innerHTML = `
+      <a id="link-manage-users" href="#">Manage Users</a>
+      `
+      // display modal Manage Users
+      document.getElementById('link-manage-users').addEventListener('click', displayModalUsers)
+    }
+  }
+}
+
+export const setup = (res) => {
+  
+  if (res.admin === 'true') {
+    const plusBlock = document.getElementById('block-plus')
+    plusBlock.addEventListener('click', () => displayModal())
+  } else {
+    const blockTitleIcon = document.getElementsByClassName('block-title-icon')
+    Array.from(blockTitleIcon).forEach(e => {
+      e.style.paddingTop = "3%"
+      e.style.marginTop = "3%"
+    })
+  }
+}
